Fix RSS template, update archetypes
THe previous RSS archetype was only displaying links to the blogs index page and the presentations index page.
This change makes it so it shows all posts in the top level index.

The change also adds the presentations archetype so the setup of the presentations is a lot quicker.
Finally the summary field is added by default to the default template to ensure it is formatted nicely on the page